+++ Sources/Sentry/SentrySDK.m
@@ -16,6 +16,7 @@
 #import "SentryLog.h"
 #import "SentryLogC.h"
 #import "SentryMeta.h"
+#import "SentryNSProcessInfoWrapper.h"
 #import "SentryOptions+Private.h"
 #import "SentryProfilingConditionals.h"
 #import "SentryReplayApi.h"
@@ -39,6 +40,8 @@
 #    import "SentryProfiler+Private.h"
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED
 
+NSString *const SENTRY_XCODE_PREVIEW_ENVIRONMENT_KEY = @"XCODE_RUNNING_FOR_PREVIEWS";
+
 @interface SentrySDK ()
 
 @property (class) SentryHub *currentHub;
@@ -47,7 +50,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 @implementation SentrySDK
-
 static SentryHub *_Nullable currentHub;
 static NSObject *currentHubLock;
 static BOOL crashedLastRunCalled;
@@ -201,6 +203,11 @@ static NSDate *_Nullable startTimestamp = nil;
 
 + (void)startWithOptions:(SentryOptions *)options
 {
+    if ([SentryDependencyContainer.sharedInstance.processInfoWrapper
+                .environment[SENTRY_XCODE_PREVIEW_ENVIRONMENT_KEY] isEqualToString:@"1"]) {
+        return;
+    }
+
     startOption = options;
     [SentryLog configure:options.debug diagnosticLevel:options.diagnosticLevel];
 
@@ -612,6 +619,21 @@ static NSDate *_Nullable startTimestamp = nil;
     [SentryContinuousProfiler stop];
 }
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED
+
+#if TEST || TESTCI
+static NSDictionary<NSString *, NSString *> *_processInfoEnvironment;
+
++ (void)setProcessInfoEnvironment:(NSDictionary<NSString *, NSString *> *)dictionary
+{
+    _processInfoEnvironment = dictionary;
+}
+
++ (NSDictionary<NSString *, NSString *> *)processInfoEnvironment
+{
+    return _processInfoEnvironment;
+}
+
+#endif
 
 @end