+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Improvements
 
 - Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
-- Add file IO tracking wrapper for Swift.Data (#4546)
+- Add file IO tracking wrapper for Swift.Data (#4605)
 
 ## 8.42.0-beta.2